pyproject.toml
[tool.poetry]
name = "jptime"
version = "1.1.0"
description = "handle japanese time format"
readme = "README.md"
authors = ["kitagawa-hr <kitagawahr@gmail.com>"]
homepage = "https://github.com/kitagawa-hr/jptime"
repository = "https://github.com/kitagawa-hr/jptime"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Typing :: Typed",
]
packages = [
    { include = "jptime" },
    { include = "jptime/py.typed" }
]

[tool.poetry.dependencies]
python = "^3.6"
python-dateutil = "^2.8.1"
japanese-number-converter = "1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.7.9"
mypy = "^0.761"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
